* added x86_64 libs and dlls for cppunit * removed warnings uncovered by switching to 64 bit target, mainly involving a different-sized size_t * decreased warning level on pcre project

+++ Core/SoarKernel/include/agent.h
@@ -546,7 +546,7 @@ kernel time and total_cpu_time greater than the derived total CPU time. REW */
   /* ----------------------- Memory utilities -------------------------- */
   
   /* Counters for memory usage of various types */
-  unsigned long       memory_for_usage[NUM_MEM_USAGE_CODES];
+  size_t       memory_for_usage[NUM_MEM_USAGE_CODES];
   
   /* List of all memory pools being used */
   memory_pool       * memory_pools_in_use;
